+++ src/Stream/LazyStream.php
@@ -59,7 +59,7 @@ final class LazyStream implements Readable
             // this trick allows to automatically close the opened files on the
             // system in order to avoid a fatal error when too many files are
             // opened. This is possible because of the rewind done in
-            // Adapter\Chunk\Fixed::__invoke() after persisting a file.
+            // File\Content\OfStream::chunks() after persisting a file.
             // This does not break be behaviour of the streams as once the stream
             // is manually closed we won't reopen it here
             /** @var Either<PositionNotSeekable, Stream> */